[audio] Add Device and DeviceWatcher
DeviceWatcher watches for ADR notifications that devices have been added
or removed. For each new device, we create a Device object which is
responsible for creating the device's pipeline and updating the
RouteGraph when the device is plugged or unplugged.

Due to the complexity of APIs used by these classes (ADR and the Mixer
service), unit tests with stubbed dependencies would be very complex.
Hence I currently plan to rely on integration tests to test these
classes.

Bug: 98652
Change-Id: Ic601ed10f663e9a28f4d34d7964f069d25cb7fbd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/787877
Reviewed-by: Alper Gungormusler <alperg@google.com>
Reviewed-by: Martin Puryear <mpuryear@google.com>
Commit-Queue: Tom Bergan <tombergan@google.com>