[i18n][test] Remove extra newlines in intl_wisdom_server
There are extra blank line in the flaky test result. The blank line
should be removed by the stdout to log forwarder. We are not not
sure the root cause. In this CL, we remove extra newlines in
intl_wisdom_server to avoid the unexpected blank line.

We also add `forward_stdout_to: "log"` to intl_wisdom_server cml.

Bug: 120366
Change-Id: I2c84d3b2a004327b50a237f6a77d5ad93819204a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/793102
Reviewed-by: Filip Filmar <fmil@google.com>
Commit-Queue: Jianpeng Chao <chaopeng@google.com>