[base-resolver] Use MAC to authenticate contexts
This allows the base-resolver to trust that it created the incoming
contexts and therefore that the contexts refer to base packages
or subpackages of base packages (i.e. it prevents clients from
forging contexts in an attempt to make base-resolver resolve a
subpackage of a non-base package).

Bug: 119496
Change-Id: Ifc340e4e9745ca9261b6c3d070f2b6c166adccd4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/792548
Commit-Queue: Ben Keller <galbanum@google.com>
Reviewed-by: Dan Johnson <computerdruid@google.com>
Reviewed-by: Sen Jiang <senj@google.com>