[magma] Copy ZirconPlatformConnection to sys_driver_cpp
The MSD C++ API has a new way of sending notifications from the driver
to sys_driver that will require a redesign of the interface to
ZirconPlatformConnection. Since sys_driver_cpp is allowed to depend on
zircon types, it makes sense to copy ZirconPlatformConnection into it
(renaming it to ZirconConnection) to make the process of redesigning the
interface simpler.

In the future this will also help reduce the complexity of some of the
lifetime management, since MagmaSystemConnection will be able to talk
more directly about the async loop and how messages on that are handled.

Change-Id: I3c141caf4b2e1f2dc3548db2d87346afafae03e7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/793132
Reviewed-by: Craig Stout <cstout@google.com>
Commit-Queue: John Bauman <jbauman@google.com>