[alc5663] Migrate acl5663-test  to safer teardown
This CL migrates acl5663-test to use a single-threaded
dispatcher loop. The motivation for this change is that
|component::OutgoingDirectory| is thread-unsafe and should
not reference a multi-threaded dispatcher. Doing so will
often lead to flakes. If the |component::OutgoingDirectory|
is being destroyed, _while_ there's an in-flight FIDL request
on the internal VFS object, there's a risk for use-after-free
failures.

The pattern used here, with |PerformBlockingWork| is safe and
compatible with multi-threaded code.

Bug: 118385
Change-Id: I321072372660a22457aed62dca9c8d8001887b6d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/791473
Commit-Queue: Gary Bressler <geb@google.com>
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Reviewed-by: Jocelyn Dang <jocelyndang@google.com>