[ti-tca6408a] Migrate ti-tca6408a to safer teardown
This CL migrates ti-tca6408a-test to use a single-threaded
dispatcher loop. The motivation for this change is that
|component::OutgoingDirectory| is thread-unsafe and should
not reference a multi-threaded dispatcher. Doing so will
often lead to flakes. If the |component::OutgoingDirectory|
is being destroyed, _while_ there's an in-flight FIDL request
on the internal VFS object, there's a risk for use-after-free
failures.

The pattern used here, with |PerformBlockingWork| is safe and
compatible with multi-threaded code.

Bug: 118385
Change-Id: I0a822afee97e3be40ce65c36d39de6bd9accef90
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/791474
Reviewed-by: Sarah Chan <spqchan@google.com>
Reviewed-by: Sam Hansen <hansens@google.com>
Commit-Queue: Gary Bressler <geb@google.com>