[async_patterns] DispatcherBound
This is the start of a new library (location up for suggestions) that
captures discussions with many folks about how to write async C++
safely. See README.md for details.

A `DispatcherBound<T>` is a helper to manage another object which
must live on a specific dispatcher. It lets one run tasks on that
dispatcher asynchronously from other execution contexts.

The primary purpose is to support types such as
|component::OutgoingDirectory| which must be destroyed on the dispatcher
thread. We have quite a few existing usages lying around that don't
destroy the directory on the right thread, and it's not trivial to
refactor them to all live on the same dispatcher. The DispatcherBound<T>
will implement the right destruction threading by posting tasks.
One just need to wrap their thread-unsafe type inside, and all accesses
including creation and destruction will happen on the dispatcher.

This is somewhat similar to the SequenceBound [1] type found in Chromium.

[1]:
https://chromium.googlesource.com/chromium/src/base/+/refs/heads/main/threading/sequence_bound.h

Bug: 119639

Change-Id: Ib0254b9787eeb3aabf986ac0f480a2b13f033b8f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/788719
Fuchsia-Auto-Submit: Yifei Teng <yifeit@google.com>
Reviewed-by: Adam Barth <abarth@google.com>
Commit-Queue: Yifei Teng <yifeit@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>