[starnix] Implement CLONE_VFORK and vfork()
Previously vfork() did a normal fork without waiting for the forked
process and the CLONE_FORK parameter to clone() was disallowed.

This implements the "vfork" feature where the calling thread is blocked
until the forked thread either calls execve() or exits().

Exit code setting was moved to a new helper function on the Task,
previously the exit code member variable was set directly in several
places. This gives a central place to unblock any vfork waiters when the
task terminates.

Bug: 114813

Change-Id: Idf44a7ebbdd9020140d2157901898367471af85e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/789727
Commit-Queue: Brett Wilson <brettw@google.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>