[session] Deflake session-manager-integration-test
This test was failing because the graphical-session was trying to
connect to the Scenic protocol but it was not routed.

The test doesn't verify any session behavior, only that it was
launched, so launch the simpler hello-world-session instead.

Fixed: 120541
Change-Id: Idfc09c4cc471dab766b0fec53d27ab91fcf5c0cf
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/794077
Reviewed-by: Xyan Bhatnagar <xbhatnag@google.com>
Fuchsia-Auto-Submit: Yegor Pomortsev <ypomortsev@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>