[bazel_sdk] merge the starlark rules with the generated SDK
The fx build build/bazel:generate_fuchsia_sdk_repository command is executed by infrastructure to build the bazel sdk and release it to CIPD. The current release does not include the starlark rules and so cannot be used by OOT customers. This change includes the fuchsia and cipd rules so that we can consume this in downstream repos.

This changes is a temporary step to unblocking the release of the fuchsia_sdk rules. The longer term solution will be to invert this relationship and have the generated rules get copied into rules_fuchsia directory. The end result should still have the same shape.

BUG: 114228
Change-Id: I4725b79baabdb0e7078fce924840d6efe11e224c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/794090
Reviewed-by: Jiaming Li <lijiaming@google.com>
Fuchsia-Auto-Submit: Chase Latta <chaselatta@google.com>
Reviewed-by: David Turner <digit@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>