[netstack] Hold stack-wide cache lock during SendMsgPreflight
Currently the implementation of SendMsgPreflight only holds the
netstack-wide destination cache lock while it duplicates the associated
eventpair for the client, and not while it is performing preflight
validation. Because of this, a TOCTTOU race is possible:

- The client calls SendMsgPreflight, which validates that a particular
  destination is valid through a given interface
- Meanwhile, elsewhere, the netstack removes that interface, takes the
  stack-wide cache lock, and resets the cache
- Back in SendMsgPreflight, the netstack duplicates the (now-reset)
  eventpair and returns it to the client, which now has a valid
  eventpair for a route through an interface that has been removed from
  the stack.

To prevent this, hold the lock on the stack-wide cache for the duration
of SendMsgPreflight. Also, change the cache lock to an RWMutex to avoid
contention between concurrent SendMsgPreflight calls, as a preflight
check only requires read access to the cache.

Fixed: 119101
Multiply: netstack-interfaces-fidl-test
Change-Id: Idb89035c160aca450ebadb5a7f24eba719685a81
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/793683
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Commit-Queue: Peter Johnston <peterjohnston@google.com>
Fuchsia-Auto-Submit: Peter Johnston <peterjohnston@google.com>
Reviewed-by: Nick Brown <nickbrow@google.com>