[ffx-test] Remove timeout to check ffx network problems.
We added this timeout to check network problems with ffx but this makes code complicated and error prone. Currently due to this timeout tests
are killed after 1 minute. To make the functionality correct we will
need to pass this timeout everywhere where we make overnet calls which is not feasible and will make the code more complex.

So remove the timeout and only check it once at the start. If we face
network issues with ffx, ffx should handle them internally.

BUG=120384

Change-Id: Ia51718d79753b1401a6233c34637c51e233da07a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/793762
Commit-Queue: Ankur Mittal <anmittal@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>
Fuchsia-Auto-Submit: Ankur Mittal <anmittal@google.com>