[starnix] Ignore CLONE_VM when used with CLONE_FORK.
We do not currently have the ability to share address spaces between
processes. This is required to implement clone(CLONE_VM) and the address
space cloning behavior specified by vfork().

By ignoring CLONE_VM when used without CLONE_THREAD (i.e. when making a
child process), this instead performs a copy-on-write clone like fork()
does. This implementation is explicitly permitted by the POSIX
specification as called out in the vfork() man page and the
implementations we have seen that use CLONE_VM are all OK with this
behavior. This flag is mainly used to avoid creating a copy of the
address space when doing a fork because it will be immediately replaced
by an exec() call.

Bug: 114813
Change-Id: I138242cbc8eb96f944ba28128e4b8805583bc198
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/794674
Reviewed-by: Theodore Dubois <tbodt@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>