[test][core] Fix a program assertion which should be a test assertion.
Fix an assert written in a test as ZX_DEBUG_ASSERT instead of using
one of the test framework's ASSERT macros (stuff like ASSERT_OK,
ASSERT_EQ, etc).

The statement being asserted here is not a program invariant (iow -
this is not something entirely under the test program's control) and
therefore should not result in termination of the test program if the
assert fails.

When this assert _does_ currently fail, it causes the core test suite
to terminate early, failing to run the remaining tests in the suite,
and potentially wedging the CI/CQ test bots until they timeout (10 min
later)

Change-Id: I810f47ef89e9546cf8184e4d30e573e3ca659b93
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/794357
Commit-Queue: John Grossman <johngro@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>