[a11y][test] Migrates to the new AddLocalChild API
The migration is using the same idea as in fxr/794259.

A fxl::WeakPtr is created in the test fixture, and initialized in
the LocalComponentImpl so that the test fixture can reach into
the LocalComponentImpl's state.

The appropriate init code was changed to wire this through, and
some waits have been set to account for lazy initialization of
the local component.

Bug: 120923
Change-Id: Ie952aff5fb3868dabe6dc4d3f55d13e6261bf6d7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/793578
Fuchsia-Auto-Submit: Jaeheon Yi <jaeheon@google.com>
Fuchsia-Auto-Submit: Filip Filmar <fmil@google.com>
Commit-Queue: Filip Filmar <fmil@google.com>
Reviewed-by: Jaeheon Yi <jaeheon@google.com>