[ffx][fho] `Arc` the `Injector` instead of `Box`
Even with the Injector boxed, its lifetime is pretty heavily
constrained, preventing any kind of deferred action on things like
setting up proxies.

This puts it in an Arc so that we can hold on to it for a bit longer and
add a `Deferred` handler (in an upcoming change list) that allows tools
to make decisions about what parts of their initialization struct they
actually need.

Bug: 120550
Bug: 120881
Change-Id: I844d1d9a4dfec0f88159b41bd788d225873cd030
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/794440
Reviewed-by: Andrew Davies <awdavies@google.com>
Commit-Queue: Megan Batty <mgnb@google.com>