[netstack] Improve interface watcher logging
A number of the panics and some of the recent logging that was added
are printing addresses as bytes rather than in their human-friendly
forms. Correct all of these, as that might also interfere with
anonymization.

Print all valid-until and preferred-until values as a duration since
boot (which is the monotonic clock's epoch).

Add tests for the Stringer impls.

Bug: 119988
Test: fx test --test-filter '*String*' netstack-gotest
Change-Id: Ia165875d7fabad558a296f570f896a222d542549
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/778775
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>
Reviewed-by: Tamir Duberstein <tamird@google.com>
Commit-Queue: Tony Gong <gongt@google.com>