[netstack3] Make ethernet device ID numbers unique
Ethernet device IDs reference state via internal shared pointer. The
numeric values are almost entirely unused, and are completely invisible
to external code. Now that they are not used as indices into an IdMap
(and DeviceId is a shared pointer to the device state), and are
therefore only useful for debugging, make them unique over the lifetime
of the netstack by keeping a stack-wide counter and incrementing it for
each insertion. This trades off a slightly longer device removal time
(since removal now requires a linear scan) for ease of debugging.

Change-Id: I2ee65c93fe3dcd523da4da8ff2aea028d5a2aaf9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/794002
Commit-Queue: Alex Konradi <akonradi@google.com>
Reviewed-by: Zeling Feng <zeling@google.com>
Reviewed-by: Nick Brown <nickbrow@google.com>