[reachability] Check for DNS activity in eventloop
Query a known domain "www.gstatic.com" with a cadence of 300
seconds to determine whether DNS is active on the system.
Confirm that the response is not none and is a non-empty
vector. Will be used in a future CL to derive DNS activity.

Test success of the lookup_ip call and each of the possible
error types (fidl error, timeout error, lookup error).

Bug: 119352
Change-Id: I918037848dabc23807f6efbadba30d6c0d4ab7f2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/793917
Reviewed-by: Jeff Martin <martinjeffrey@google.com>
Commit-Queue: Antonio Linhart <antoniolinhart@google.com>
Reviewed-by: Charles Celerier <chcl@google.com>