[debugger] Fix number separator parsing.
In Rust code, expressions with "__0" in it would get confusing parsings.
It turns out we were parsing the string as a floating-point number
because '_' is a digit separator in Rust.

Disallows digit separators (also for C which uses ') in the first
position. Add tests.

Change-Id: I3df5e22da33e7f75d10c4f97793663010f55b4a9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/795281
Reviewed-by: Jacob Rutherford <jruthe@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>