[starnix] Introduce abstraction for file I/Os
This CL introduce the notion of InputBuffer and OutputBuffer for file
read and writes.

This simplifies tests, as they do not need to create memory mapping to
execute read and write on files.

This will also be used to write a FIDL file server for starnix
FileObject.

Change-Id: I15b3db4fa9c7ba8074c80d60cf71052fc4feade1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/793974
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Fuchsia-Auto-Submit: Benjamin Lerman <qsr@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>