[storage][perf] Add --load-blobs-for-tracing flag
When trying to collect a trace immediately after modifying a filesystem
or a benchmark, the start of the trace will be polluted with downloading
the new blobs, writing the blobs to blobfs, and then paging the blobs
back in. Running the benchmark component once with the
--load-blobs-for-tracing flag before running it again with tracing
enabled will remove most of the blob loading from the start of the
trace.

Change-Id: Id835da636ce8df6725ebd6cd5787177e836d8e4d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/795126
Reviewed-by: Chris Tam <godtamit@google.com>
Commit-Queue: Chris Drouillard <cdrllrd@google.com>