[cmx] Delete tiles_ctl.
This tool relies on access to `/hub` which was removed from the
system a couple of months ago. Since the tool doesn't work on ToT
and we haven't heard from any developers regarding the breakage,
it seems safe to assume that it is no longer in use.

Bug: 100732
Change-Id: I8019f8842d0c3366cec5e46d75b68151f59a6a58
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/794976
Fuchsia-Auto-Submit: Adam Perry <adamperry@google.com>
Reviewed-by: David Worsham <dworsham@google.com>
Reviewed-by: Shai Barack <shayba@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Gabe Schine <thatguy@google.com>