[netstack3] Access device addr status via context
Add a new trait for accessing IP per-device state and use it to
implement the existing address assignment state lookup. This is a no-op
for now but will make it possible to change all &SC arguments in context
and handler traits to &mut SC in a future CL.

This change is necessary to make with_devices take &mut self. Without
this, the impls of with_address_statuses will attempt to pass the same
&mut self into the closure that is an argument to with_devices, which
the compiler forbids.

Bug: 120973
Change-Id: I07ce5f15b5b358ea6770aae2156ce1c594780dfd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/796122
Reviewed-by: Zeling Feng <zeling@google.com>
Commit-Queue: Alex Konradi <akonradi@google.com>
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>