[starnix] Allow clone3 to specify stack.
This CL fixes a bug in starnix's clone3 implementation that didn't allow
it to use a different stack.

In the regular `clone` system call, the user has to specify the topmost
byte of the stack. In `clone3`, the user has to specify the *lowest*
byte of the stack, as well as the stack size. There was a bug in the
clone3 implementation, that it didn't take the size into account.

The new implementation of `do_clone` now sets the stack to the provided
address + the size (note that in the clone case, size will be 0).

Co-authored-by: mvanotti@google.com
Test: fx test syscalls_test
Bug: 120202
Change-Id: I31a12758e1a450228c2cf058fe2bad2ae28ff17b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/794262
Commit-Queue: Marina Ciocea <marinaciocea@google.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Reviewed-by: Brett Wilson <brettw@google.com>
Reviewed-by: Marco Vanotti <mvanotti@google.com>