[gnsdk-fidl] Add new C++ bindings to fidl_library
This adds support for generating and compiling the new c++ bindings
with the GN SDK. The names of the new targets mirror the names used
in-tree.

HLCPP bindings are included with the _hlcpp suffix, and they are also
aliased to the target without a suffix until all usages can be migrated.

Bug: 115033
Change-Id: Icfb6e28baa0c0210ac67a1ecd68e0f1be97d62e8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/793915
Reviewed-by: Sébastien Marchand <sebmarchand@google.com>
Reviewed-by: Yifei Teng <yifeit@google.com>
Commit-Queue: Bryant Chandler <bryantchandler@google.com>