[bt][avrcp-integration-tests] Retry channel reads until success
The underlying fasync::Socket sometimes returns a successful read of 0
bytes when using read_datagram() to read on the channel. Keep retrying
the read until a nonzero buffer is read, or an error occurs in which
case the test will exit with failure.

Bug: 120232
Test: fx test bt-avrcp-integration-tests
Multiply: bt-avrcp-integration-tests
Change-Id: I3a25768e0ed25a96d47fdfdbffce677150c3e427
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/796442
Reviewed-by: Joshua Conner <joshconner@google.com>
Commit-Queue: Ani Ramakrishnan <aniramakri@google.com>